Provide system information (task #2450)
* Introduced System controller for system-related actions
* Added /system/info for display of the system information

The view itself is broken down into a variety of elements, each
of each gets the relevant information.  This helps with a potential
need to display parts of the system information in different areas
of the application.

The only thing pending is to refactor project name, project URL,
and project versions, but this change needs an upstream update to
how we handle these variables.