TravisCI and Bitbucket Piplelines Validator fix
Validator should run before app:install so that it can check if
the configuration and migrations are good to go.  But for the
validator to run (as a CakePHP plugin shell), we need CakePHP
to be operational.  For that, we need the `.env` file to be in
place.

So, with this fix, we create the `.env` file prior to the attempt
to install the application (which will run the migrations, which
are possibly broken).  Once the `.env` file is in place, we can
run the validator, and if it passes, then we can run the phake's
`app:install` which will trigger migrations.