Warn instead of fail on chmod/chown/chgrp
Permission and ownership changing operations are very basic and
rarely fail in real environments.  However, these are often
disabled or limited in testing environments, like TravisCI and
BitBucket Pipelines.

With this fix, instead of failing the build, we'll assume that the
application can still run just fine, and simply emit a warning
message when chmod/chown/chgrp operation fail.