Added debug_kit database connection
DebugKit plugin defines its own database connection, which,
if you don't specify anything, uses Sqlite database in the tmp
folder.

Since DebugKit plugin is loaded conditionally, depending on the
value of the DEBUG variable, its configuration might not always
be loaded, and hence the connection is not defined.  Usually,
this works just fine, but in some cases, it breaks the process
of deployment, which runs migrations of all plugins.

To avoid the issue, debug_kit configurationw as copied verbatim
from DebugKit's bootstrap.php into the application's config/app.php
file.