Test database re-creation
During phake app:install we currently create the test database and
we run the migrations on it to verify that they run successfully.

After the installation is done, unit tests are run and will fail if
a fixture table already has a foreign key constraints with another
table which exists in the test database from the migrations.

To avoid this issue, after we verify that migrations are run successfully
on the test database we drop the database and re-create it without
any tables.