Optional disable of outgoing emails sending (task #4458)
Control whether or not the outgoing emails will be sent out via
the `EMAIL_ENABLED=1` setting in `.env`.  By default, the emails
are sent out.

If the email sending is disabled, we switch to the Debug email
transport of CakePHP, which will avoid sendign emails.  If sending
is enabled, then we use either the Smtp transport (if `SMTP_ENABLED`
is true in `.env`), or fallback on the Mail transport, which uses
PHP built-in `mail()` functionality.