Updated bootstrap.php and EmailShell for disabled email (task #4458)
Firstly, we need to `Configured::read()` rather than
`Configure::consume()` the email configuration.  Consume deletes
the configuration values after reading, but we need them further
down in the application.

Secondly, updated the EmailShell to check for whether or not the
email sending is enabled.  Otherwise it would always return success
via the Debug transport.