Added experimental About section to Project info (task #5473)
Project information tab looks quite empty, so we an About section
should fill it up a bit.  We still need to find a more elegant way
for this section to include information about inheritted projects,
links, versions, and logos, but it should for now.

At some point later, the About section should also be available as
a separate System's controller action.  Without any sensitive
information on it (unlike the System Info), it can be linked to
from the footer version number and other locations.