Random salt (task #5518)
* Removed SALT from .env.example
* Replaced getting the SALT from the environment with the call
  to the `Qobo\Utils\Utility\Salt::getSalt()`.

NOTE: This change requires a version of cakephp-utils plugin with
      the following PR merged in:
      https://github.com/QoboLtd/cakephp-utils/pull/103

The security salt is now generated on the fly and is stored in the
`tmp/security.salt` file.  If the file does not exist or contains
an invalid salt (too short of a string, for example), the file
will be re-generated with the proper salt string.

For all affected users (invalid tokens, CSRF headers, etc), you'll
need to clear cookies, reload the page, and try again.