Remove legacy Google Analytics integration
A long time ago in a galaxy far, far away, we used to embed
Google Analytics code snippet if the UA account was set in
the environment configuration.

A lot of things changed since that time.  In particular, the
move to AdminLTE dropped off the snippet and it was never used
since.

This commit removes this old, unused, and unnecessary support, as
there are better ways to do this.