Minor fix of the page title
If the controller action does set the title of its own, we use
the name of the controller as is.  With this change we pass the
controller name through the Inflector's underscore and humanize
methods, which produces a nicer result for contnroller names that
use multiple words, like TradingAccounts (becomes: Trading Accounts).